move ColumnViewDelegate to own file, because it can be used also by the TreeView
